+++ src/frontend/js/components/Root/index.tsx
@@ -10,6 +10,7 @@ import ErrorBoundary from 'utils/errors/ErrorBoundary';
 import context from 'utils/context';
 
 const CourseProductItem = lazy(() => import('components/CourseProductItem'));
+const CourseAddToWishlist = lazy(() => import('components/CourseAddToWishlist'));
 const CourseRunEnrollment = lazy(() => import('components/CourseRunEnrollment'));
 const LanguageSelector = lazy(() => import('components/LanguageSelector'));
 const LtiConsumer = lazy(() => import('components/LtiConsumer'));
@@ -23,6 +24,7 @@ const Dashboard = lazy(() => import('components/Dashboard'));
 // for type-safety when we call them. This will let us use the props for any top-level component in a
 // way TypeScript understand and accepts
 interface ComponentLibrary {
+  CourseAddToWishlist: typeof CourseAddToWishlist;
   CourseProductItem: typeof CourseProductItem;
   CourseRunEnrollment: typeof CourseRunEnrollment;
   Dashboard: typeof Dashboard;
@@ -35,6 +37,7 @@ interface ComponentLibrary {
 }
 // Actually create the component map that we'll use below to access our component classes
 const componentLibrary: ComponentLibrary = {
+  CourseAddToWishlist,
   CourseProductItem,
   CourseRunEnrollment,
   Dashboard,